+++ src/Paket.Core/Nuspec.fs
@@ -22,10 +22,11 @@ type Nuspec =
     { References : NuspecReferences 
       Dependencies : (PackageName * VersionRequirement * FrameworkRestrictions) list
       OfficialName : string
+      LicenseUrl : string
       FrameworkAssemblyReferences : FrameworkAssemblyReference list }
 
-    static member All = { References = NuspecReferences.All; Dependencies = []; FrameworkAssemblyReferences = []; OfficialName = "" }
-    static member Explicit references = { References = NuspecReferences.Explicit references; Dependencies = []; FrameworkAssemblyReferences = []; OfficialName = "" }
+    static member All = { References = NuspecReferences.All; Dependencies = []; FrameworkAssemblyReferences = []; OfficialName = ""; LicenseUrl = "" }
+    static member Explicit references = { References = NuspecReferences.Explicit references; Dependencies = []; FrameworkAssemblyReferences = []; OfficialName = ""; LicenseUrl = "" }
     static member Load(fileName : string) = 
         let fi = FileInfo(fileName)
         if not fi.Exists then Nuspec.All
@@ -76,6 +77,11 @@ type Nuspec =
                 |> List.map dependency
                 |> List.append frameworks
                 |> Requirements.optimizeRestrictions 
+
+            let licenseUrl =                 
+                match doc |> getNode "package" |> optGetNode "metadata" |> optGetNode "licenseUrl" with
+                | Some link -> link.InnerText
+                | None -> ""     
             
             let references = 
                 doc
@@ -114,4 +120,5 @@ type Nuspec =
             { References = if references = [] then NuspecReferences.All else NuspecReferences.Explicit references
               Dependencies = dependencies
               OfficialName = officialName
+              LicenseUrl = licenseUrl
               FrameworkAssemblyReferences = frameworkAssemblyReferences }