+++ src/Paket.Core/PackageResolver.fs
@@ -45,6 +45,7 @@ type PackageDetails =
     { Name : PackageName
       Source : PackageSource
       DownloadLink : string
+      LicenseUrl : string
       Unlisted : bool
       DirectDependencies : DependencySet }
 
@@ -53,7 +54,7 @@ type ResolvedPackage =
     { Name : PackageName
       Version : SemVerInfo
       Dependencies : DependencySet
-      Unlisted : bool 
+      Unlisted : bool
       Settings : InstallSettings
       Source : PackageSource }
 
@@ -75,14 +76,14 @@ let cleanupNames (model : PackageResolution) : PackageResolution =
                  Map.add (NormalizedPackageName package.Name) cleanup map) Map.empty
 
 [<RequireQualifiedAccess>]
-type ResolvedPackages =
+type Resolution =
 | Ok of PackageResolution
 | Conflict of Set<PackageRequirement> * Set<PackageRequirement>
     with
     member this.GetModelOrFail() =
         match this with
-        | ResolvedPackages.Ok model -> model
-        | ResolvedPackages.Conflict(closed,stillOpen) ->
+        | Resolution.Ok model -> model
+        | Resolution.Conflict(closed,stillOpen) ->
 
             let errorText = ref ""
 
@@ -141,7 +142,7 @@ let calcOpenRequirements (exploredPackage:ResolvedPackage,globalFrameworkRestric
     |> Set.union rest
 
 type Resolved = {
-    ResolvedPackages : ResolvedPackages
+    ResolvedPackages : Resolution
     ResolvedSourceFiles : ModuleResolver.ResolvedSourceFile list }
 
 /// Resolves all direct and transitive dependencies
@@ -195,9 +196,9 @@ let Resolve(getVersionsF, getPackageDetailsF, globalFrameworkRestrictions, rootD
             versions
         | true,versions -> versions        
 
-    let rec improveModel (filteredVersions:Map<PackageName, (SemVerInfo list * bool)>,packages:ResolvedPackage list,closed:Set<PackageRequirement>,stillOpen:Set<PackageRequirement>) =
-        if Set.isEmpty stillOpen then
-            // we're done. check if we have a valid resolution and return it
+    let rec step (filteredVersions:Map<PackageName, (SemVerInfo list * bool)>,selectedPackageVersions:ResolvedPackage list,closedRequirements:Set<PackageRequirement>,openRequirements:Set<PackageRequirement>) =
+        if Set.isEmpty openRequirements then
+            // we're done. re-check if we have a valid resolution and return it
             let isOk =
                 filteredVersions
                 |> Map.forall (fun _ v ->
@@ -207,20 +208,20 @@ let Resolve(getVersionsF, getPackageDetailsF, globalFrameworkRestrictions, rootD
 
             if isOk then
                 let resolution =
-                    packages 
+                    selectedPackageVersions 
                     |> Seq.fold (fun map p -> Map.add (NormalizedPackageName p.Name) p map) Map.empty
 
-                ResolvedPackages.Ok(resolution)
+                Resolution.Ok(resolution)
             else
-                ResolvedPackages.Conflict(closed,stillOpen)
+                Resolution.Conflict(closedRequirements,openRequirements)
         else
-            let packageCount = packages |> List.length
-            verbosefn "  %d packages in resolution. %d requirements left" packageCount stillOpen.Count
+            let packageCount = selectedPackageVersions |> List.length
+            verbosefn "  %d packages in resolution. %d requirements left" packageCount openRequirements.Count
             
-            let dependency =
-                let currentMin = ref (Seq.head stillOpen)
+            let currentRequirement =
+                let currentMin = ref (Seq.head openRequirements)
                 let currentBoost = ref 0
-                for d in stillOpen do
+                for d in openRequirements do
                     let boost = 
                         match conflictHistory.TryGetValue(NormalizedPackageName d.Name) with
                         | true,c -> -c
@@ -229,64 +230,66 @@ let Resolve(getVersionsF, getPackageDetailsF, globalFrameworkRestrictions, rootD
                         currentMin := d
                         currentBoost := boost
                 !currentMin
-            verbosefn "  Trying to resolve %O" dependency
+            verbosefn "  Trying to resolve %O" currentRequirement
 
-            let allVersions = ref []
+            let availableVersions = ref []
             let compatibleVersions = ref []
             let globalOverride = ref false
      
-            match Map.tryFind dependency.Name filteredVersions with
+            match Map.tryFind currentRequirement.Name filteredVersions with
             | None ->
-                allVersions := getAllVersions(dependency.Sources,dependency.Name,dependency.VersionRequirement.Range)
-                compatibleVersions := List.filter dependency.VersionRequirement.IsInRange (!allVersions)
-                if dependency.VersionRequirement.Range.IsGlobalOverride then
+                // we didn't select a version yet so all versions are possible
+                availableVersions := getAllVersions(currentRequirement.Sources,currentRequirement.Name,currentRequirement.VersionRequirement.Range)
+                compatibleVersions := List.filter currentRequirement.VersionRequirement.IsInRange (!availableVersions)
+                if currentRequirement.VersionRequirement.Range.IsGlobalOverride then
                     globalOverride := true
                 else
                     if !compatibleVersions = [] then
-                        let prereleases = List.filter (dependency.IncludingPrereleases().VersionRequirement.IsInRange) (!allVersions)
+                        let prereleases = List.filter (currentRequirement.IncludingPrereleases().VersionRequirement.IsInRange) (!availableVersions)
                         if allPrereleases prereleases then
-                            allVersions := prereleases
+                            availableVersions := prereleases
                             compatibleVersions := prereleases
             | Some(versions,globalOverride') -> 
+                // we already selected a version so we can't pick a different
                 globalOverride := globalOverride'
-                allVersions := versions
+                availableVersions := versions
                 if globalOverride' then
                     compatibleVersions := versions
                 else
                     compatibleVersions := 
-                        List.filter (fun v -> dependency.VersionRequirement.IsInRange(v,dependency.Parent.IsRootRequirement() |> not)) versions
+                        List.filter (fun v -> currentRequirement.VersionRequirement.IsInRange(v,currentRequirement.Parent.IsRootRequirement() |> not)) versions
 
             if !compatibleVersions = [] then
-                if dependency.Parent.IsRootRequirement() then    
-                    let versionText = String.Join(Environment.NewLine + "     - ",List.sort !allVersions)
+                if currentRequirement.Parent.IsRootRequirement() then    
+                    let versionText = String.Join(Environment.NewLine + "     - ",List.sort !availableVersions)
                     failwithf "Could not find compatible versions for top level dependency:%s     %A%s   Available versions:%s     - %s%s   Try to relax the dependency or allow prereleases." 
-                        Environment.NewLine (dependency.ToString()) Environment.NewLine Environment.NewLine versionText Environment.NewLine
+                        Environment.NewLine (currentRequirement.ToString()) Environment.NewLine Environment.NewLine versionText Environment.NewLine
                 else
                     // boost the conflicting package, in order to solve conflicts faster
-                    match conflictHistory.TryGetValue(NormalizedPackageName dependency.Name) with
-                    | true,count -> conflictHistory.[NormalizedPackageName dependency.Name] <- count + 1
-                    | _ -> conflictHistory.Add(NormalizedPackageName dependency.Name, 1)
+                    match conflictHistory.TryGetValue(NormalizedPackageName currentRequirement.Name) with
+                    | true,count -> conflictHistory.[NormalizedPackageName currentRequirement.Name] <- count + 1
+                    | _ -> conflictHistory.Add(NormalizedPackageName currentRequirement.Name, 1)
                     
                     if verbose then
                         tracefn "  Conflicts with:"
                     
-                        closed
-                        |> Seq.filter (fun d -> d.Name = dependency.Name)
+                        closedRequirements
+                        |> Seq.filter (fun d -> d.Name = currentRequirement.Name)
                         |> fun xs -> String.Join(Environment.NewLine + "    ",xs)
                         |> tracefn "    %s"
 
-                        match filteredVersions |> Map.tryFind dependency.Name with
-                        | Some (v,_) -> tracefn "    Package %O was already pinned to %O" dependency.Name v
+                        match filteredVersions |> Map.tryFind currentRequirement.Name with
+                        | Some (v,_) -> tracefn "    Package %O was already pinned to %O" currentRequirement.Name v
                         | None -> ()
 
                         tracefn "    ==> Trying different resolution."
 
 
-            let sortedVersions =                
-                if dependency.Parent.IsRootRequirement() then
+            let sortedVersions =
+                if currentRequirement.Parent.IsRootRequirement() then
                     List.sort !compatibleVersions |> List.rev
                 else
-                    match dependency.ResolverStrategy with
+                    match currentRequirement.ResolverStrategy with
                     | ResolverStrategy.Max -> List.sort !compatibleVersions |> List.rev
                     | ResolverStrategy.Min -> List.sort !compatibleVersions
 
@@ -294,27 +297,27 @@ let Resolve(getVersionsF, getPackageDetailsF, globalFrameworkRestrictions, rootD
                 sortedVersions
                 |> List.fold (fun (allUnlisted,state) versionToExplore ->
                     match state with
-                    | ResolvedPackages.Conflict _ ->
-                        let exploredPackage = getExploredPackage(dependency,versionToExplore)
+                    | Resolution.Conflict _ ->
+                        let exploredPackage = getExploredPackage(currentRequirement,versionToExplore)
                         if exploredPackage.Unlisted && not useUnlisted then 
                             allUnlisted,state 
                         else                
-                            let newFilteredVersions = Map.add dependency.Name ([versionToExplore],!globalOverride) filteredVersions
+                            let newFilteredVersions = Map.add currentRequirement.Name ([versionToExplore],!globalOverride) filteredVersions
                         
-                            let newOpen = calcOpenRequirements(exploredPackage,globalFrameworkRestrictions,versionToExplore,dependency,closed,stillOpen)
+                            let newOpen = calcOpenRequirements(exploredPackage,globalFrameworkRestrictions,versionToExplore,currentRequirement,closedRequirements,openRequirements)
                             let newPackages =
-                                exploredPackage::(packages |> List.filter (fun p -> NormalizedPackageName p.Name <> NormalizedPackageName exploredPackage.Name || p.Version <> exploredPackage.Version))
+                                exploredPackage::(selectedPackageVersions |> List.filter (fun p -> NormalizedPackageName p.Name <> NormalizedPackageName exploredPackage.Name || p.Version <> exploredPackage.Version))
 
-                            let improved = improveModel (newFilteredVersions,newPackages,Set.add dependency closed,newOpen)                            
+                            let improved = step (newFilteredVersions,newPackages,Set.add currentRequirement closedRequirements,newOpen)                            
                             (exploredPackage.Unlisted && allUnlisted),improved
                             
-                    | ResolvedPackages.Ok _ -> allUnlisted,state)
-                        (true,ResolvedPackages.Conflict(closed,stillOpen))
+                    | Resolution.Ok _ -> allUnlisted,state)
+                        (true,Resolution.Conflict(closedRequirements,openRequirements))
             
             match tryToImprove false with
-            | true,ResolvedPackages.Conflict(_) -> tryToImprove true |> snd       
+            | true,Resolution.Conflict(_) -> tryToImprove true |> snd       
             | _,x-> x
 
-    match improveModel (Map.empty, [], Set.empty, Set.ofList rootDependencies) with
-    | ResolvedPackages.Conflict(_) as c -> c
-    | ResolvedPackages.Ok model -> ResolvedPackages.Ok(cleanupNames model)
+    match step (Map.empty, [], Set.empty, Set.ofList rootDependencies) with
+    | Resolution.Conflict(_) as c -> c
+    | Resolution.Ok model -> Resolution.Ok(cleanupNames model)