+++ tests/Paket.Tests/Resolver/ConflictGraphSpecs.fs
@@ -35,8 +35,8 @@ let defaultPackage =
 [<Test>]
 let ``should analyze graph and report conflict``() = 
     match safeResolve graph [ "A", VersionRange.AtLeast "1.0" ] with
-    | ResolvedPackages.Ok _ -> failwith "we expected an error"
-    | ResolvedPackages.Conflict(_,stillOpen) ->
+    | Resolution.Ok _ -> failwith "we expected an error"
+    | Resolution.Conflict(_,stillOpen) ->
         let conflicting = stillOpen |> Seq.head 
         conflicting.Name |> shouldEqual (PackageName "D")
         conflicting.VersionRequirement.Range |> shouldEqual (VersionRange.Exactly "1.4")
@@ -53,8 +53,8 @@ let graph2 =
 [<Test>]
 let ``should analyze graph2 and report conflict``() = 
     match safeResolve graph2 [ "A", VersionRange.AtLeast "1.0" ] with
-    | ResolvedPackages.Ok _ -> failwith "we expected an error"
-    | ResolvedPackages.Conflict(_,stillOpen) ->
+    | Resolution.Ok _ -> failwith "we expected an error"
+    | Resolution.Conflict(_,stillOpen) ->
         let conflicting = stillOpen |> Seq.head 
         conflicting.Name |> shouldEqual (PackageName "D")
         conflicting.VersionRequirement.Range |> shouldEqual (VersionRange.Between("1.4", "1.5"))