+++ RELEASE_NOTES.md
@@ -1,6 +1,15 @@
-#### 0.38.0-alpha001 - 28.03.2015
+#### 0.39.0-alpha001 - 01.04.2015
 * Upgrading to .NET 4.5
 * Removing DotNetZip and using the .NET 4.5 Zip APIs instead - https://github.com/fsprojects/Paket/pull/732
+
+#### 0.38.3 - 01.04.2015
+* Detect `sl` as Silverlight - https://github.com/fsprojects/Paket/issues/744
+
+#### 0.38.2 - 31.03.2015
+* Added frameworkAssemblies to nuspec templating - https://github.com/fsprojects/Paket/issues/740
+
+#### 0.38.0 - 30.03.2015
+* The restore process downloads package licenses automatically - https://github.com/fsprojects/Paket/pull/737
 
 #### 0.37.0 - 28.03.2015
 * Fallback to NuGet.exe if the bootstrapper fails to download from GitHub - https://github.com/fsprojects/Paket/pull/733